* master: (172 commits)
  update versions
  clean up build
  next version
  next version
  Add tests
  fix tests
  update
  trigger dependency update
  upload coverage
  upload coverage
  updates
  trigger dependency update
  add tests
  support optional x509
  clean up build
  update build config
  [skip ci]
  clean up themes
  clean up themes
  update coverage config
  ...

# Conflicts:
#	support/cas-server-support-otp-mfa-core/src/main/java/org/apereo/cas/otp/util/QRUtils.java

+++ core/cas-server-core-services/src/test/java/org/apereo/cas/services/AbstractServiceRegistryTests.java
@@ -555,7 +555,7 @@ public abstract class AbstractServiceRegistryTests {
      * @param randomId addition to service name for uniqueness.
      * @return new registered service object
      */
-    protected static AbstractRegisteredService buildRegisteredServiceInstance(final int randomId,
+    protected static AbstractRegisteredService buildRegisteredServiceInstance(final long randomId,
                                                                               final Class<? extends RegisteredService> registeredServiceClass) {
         val id = String.format("^http://www.serviceid%s.org", randomId);
         val rs = RegisteredServiceTestUtils.getRegisteredService(id, registeredServiceClass);
